+++ src/RobotStudio.Domain/Articulated/IndustrialArmJointId.cs
@@ -1,0 +1,11 @@
+namespace RobotStudio.Domain.Articulated;
+
+public enum IndustrialArmJointId
+{
+    J1Base,
+    J2Shoulder,
+    J3Elbow,
+    J4WristRoll,
+    J5WristPitch,
+    J6ToolRoll
+}